deps/v8/test/inspector/regress/regress-crbug-1253277-expected.txt

Regression test for crbug.com/1253277

Running test: test
function foo(){}|_|foo()
